Fix a bunch of ARM32 printf() warnings (Ubuntu 22.04)

+++ lib/reg/common.h
@@ -84,8 +84,8 @@ static inline time_t randomize_expires(unsigned int expires_ts)
 		expires_dur = max_expires;
 
 	ret = expires_dur + get_act_time();
-	LM_DBG("randomized expiry ts from %u to %lu (adj: %d/%d, "
-	       "max_deviation: %d)\n", expires_ts, ret, expires_adj,
+	LM_DBG("randomized expiry ts from %u to %lld (adj: %d/%d, "
+	       "max_deviation: %d)\n", expires_ts, (long long)ret, expires_adj,
 	       (int)ret - (int)expires_ts, expires_max_deviation);
 
 	return ret;
